criado enum projectStatus and movido TarefaStatusEnum para a pasta de enums

+++ src/main/java/com/bmt/webApp/model/Project.java
@@ -28,7 +28,7 @@ public class Project {
     private String descricao;
     private LocalDateTime dataInicio;
     private LocalDateTime dataFim;
-    private Double completionPercentage;
+    private Integer completionPercentage;
     private String status;//New, Em andamento e Concluído
 
     //Relacionamento 1-N com Tarefa
@@ -74,7 +74,7 @@ public class Project {
     public void setCompletionPercentage(Double completionPercentage){
         this.completionPercentage = completionPercentage;
     }
-    
+
     public String getStatus() {
         return status;
     }
